requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
openai
pinecone
python-dotenv
langdetect
pymongo
python-multipart
pydantic
requests
beautifulsoup4
pypdf
email-validator
jinja2
pycalendar
google-auth
google-auth-oauthlib
google-api-python-client
mailchimp3
hubspot-api-client
langchain-pinecone
bcrypt
cryptography
redis
tenacity